+++ pullbug/__init__.py
@@ -1,4 +1,7 @@
-# flake8: noqa
 from pullbug.github_bug import GithubBug
-from pullbug.logger import PullBugLogger
 from pullbug.messages import Messages
+
+__all__ = [
+    'GithubBug',
+    'Messages',
+]