+++ pullbug/github_bug.py
@@ -1,33 +1,33 @@
-import logging
 import os
+from typing import List, Optional, Tuple
 
+import woodchips
 from github import Github
 
-from pullbug.logger import PullBugLogger
 from pullbug.messages import Messages
 
-LOGGER = logging.getLogger(__name__)
+LOGGER_NAME = 'pullbug'
 
 
 class GithubBug:
     def __init__(
         self,
-        pulls=False,
-        issues=False,
-        github_token=None,
-        github_owner=None,
-        github_state='open',
-        github_context=None,
-        discord=False,
-        discord_url=None,
-        slack=False,
-        slack_token=None,
-        slack_channel=None,
-        rocketchat=False,
-        rocketchat_url=None,
-        repos=None,
-        drafts=False,
-        location=os.path.expanduser('~/pullbug'),
+        pulls: Optional[bool] = False,
+        issues: Optional[bool] = False,
+        github_token: Optional[str] = None,
+        github_owner: Optional[str] = None,
+        github_state: Optional[str] = 'open',
+        github_context: Optional[str] = None,
+        discord: Optional[bool] = False,
+        discord_url: Optional[str] = None,
+        slack: Optional[bool] = False,
+        slack_token: Optional[str] = None,
+        slack_channel: Optional[str] = None,
+        rocketchat: Optional[bool] = False,
+        rocketchat_url: Optional[str] = None,
+        repos: Optional[str] = None,
+        drafts: Optional[str] = False,
+        location: str = os.path.expanduser('~/pullbug'),
     ):
         # Parameter variables
         self.pulls = pulls
@@ -53,8 +53,9 @@ class GithubBug:
 
     def run(self):
         """Run the logic to get PR's from GitHub and send that data via message."""
-        PullBugLogger._setup_logging(LOGGER, self.location)
-        LOGGER.info('Running Pullbug...')
+        self.setup_logger()
+        logger = woodchips.get(LOGGER_NAME)
+        logger.info('Running Pullbug...')
         self.run_missing_checks()
 
         repos = self.get_repos()
@@ -63,7 +64,7 @@ class GithubBug:
             pull_requests = self.get_pull_requests(repos)
             if pull_requests == []:
                 message = 'No pull requests are available from GitHub.'
-                LOGGER.info(message)
+                logger.info(message)
                 # TODO: Do we want to send this message here?
             else:
                 message_preamble = (
@@ -79,7 +80,7 @@ class GithubBug:
             issues = self.get_issues(repos)
             if issues == []:
                 message = 'No issues are available from GitHub.'
-                LOGGER.info(message)
+                logger.info(message)
                 # TODO: Do we want to send this message here?
             else:
                 message_preamble = '\n:bug: *The following issues on GitHub are still open and need your help!*\n'
@@ -89,7 +90,16 @@ class GithubBug:
 
                 self.send_messages(messages, discord_messages)
 
-        LOGGER.info('Pullbug finished bugging!')
+        logger.info('Pullbug finished bugging!')
+
+    def setup_logger(self):
+        """Setup a `woodchips` logger for the project."""
+        logger = woodchips.Logger(
+            name=LOGGER_NAME,
+            level='INFO',
+        )
+        logger.log_to_console()
+        logger.log_to_file(location=os.path.join(self.location, 'logs'))
 
     def run_missing_checks(self):
         """Check that values are set based on configuration before proceeding."""
@@ -109,35 +119,44 @@ class GithubBug:
             self.throw_missing_error('rocketchat_url')
 
     @staticmethod
-    def throw_missing_error(missing_flag):
+    def throw_missing_error(missing_flag: str):
         """Raise an error based on what env variables are missing."""
+        logger = woodchips.get(LOGGER_NAME)
         message = f'No {missing_flag} set. Please correct and try again.'
-        LOGGER.critical(message)
+        logger.critical(message)
+
         raise ValueError(message)
 
     def get_repos(self):
         """Get all repos of the github_owner."""
-        LOGGER.info('Bugging GitHub for repos...')
+        logger = woodchips.get(LOGGER_NAME)
+
+        logger.info('Bugging GitHub for repos...')
+
         if self.github_context == 'orgs':
             repos = self.github_instance.get_organization(self.github_owner).get_repos()
         elif self.github_context == 'users':
             repos = self.github_instance.get_user(self.github_owner).get_repos()
         else:
             # Can't determine github_context
+            repos = []
             pass
 
         if self.repos:
             formatted_repos_list = [repo.strip() for repo in self.repos]
             repos = [repo for repo in repos if repo.name.lower() in formatted_repos_list]
 
-        LOGGER.info('GitHub repos retrieved!')
+        logger.info('GitHub repos retrieved!')
 
         return repos
 
-    def get_pull_requests(self, repos):
+    def get_pull_requests(self, repos: List) -> List:
         """Grab all pull requests from each repo and return a flat list of pull requests."""
-        LOGGER.info('Bugging GitHub for pull requests...')
+        logger = woodchips.get(LOGGER_NAME)
+
+        logger.info('Bugging GitHub for pull requests...')
         pull_requests = []
+
         for repo in repos:
             repo_pull_requests = repo.get_pulls(state=self.github_state)
             if repo_pull_requests:
@@ -148,14 +167,17 @@ class GithubBug:
 
         flat_pull_requests_list = [pull_request for pull_request in pull_requests for pull_request in pull_request]
 
-        LOGGER.info('Pull requests retrieved!')
+        logger.info('Pull requests retrieved!')
 
         return flat_pull_requests_list
 
-    def get_issues(self, repos):
+    def get_issues(self, repos: List) -> List:
         """Grab all issues from each repo and return a flat list of issues."""
-        LOGGER.info('Bugging GitHub for issues...')
+        logger = woodchips.get(LOGGER_NAME)
+
+        logger.info('Bugging GitHub for issues...')
         issues = []
+
         for repo in repos:
             repo_issues = repo.get_issues(state=self.github_state)
             if repo_issues:
@@ -166,14 +188,15 @@ class GithubBug:
 
         flat_issues_list = [issue for issue in issues for issue in issue]
 
-        LOGGER.info('Issues retrieved!')
+        logger.info('Issues retrieved!')
 
         return flat_issues_list
 
-    def iterate_pull_requests(self, pull_requests):
+    def iterate_pull_requests(self, pull_requests: List) -> Tuple[List, List]:
         """Iterate through each pull request of a repo and build the message array."""
         message_array = []
         discord_message_array = []
+
         for pull_request in pull_requests:
             if pull_request.draft and not self.drafts:
                 # Exclude drafts if the user doesn't want them included
@@ -186,10 +209,11 @@ class GithubBug:
         return message_array, discord_message_array
 
     @staticmethod
-    def iterate_issues(issues):
+    def iterate_issues(issues: List) -> Tuple[List[str], List[str]]:
         """Iterate through each issue of a repo and build the message array."""
         message_array = []
         discord_message_array = []
+
         for issues in issues:
             message, discord_message = Messages.prepare_issues_message(issues)
             message_array.append(message)
@@ -197,7 +221,10 @@ class GithubBug:
 
         return message_array, discord_message_array
 
-    def send_messages(self, messages, discord_messages):
+    def send_messages(self, messages: List[str], discord_messages: List[str]):
+        """Sends a message to the messaging platforms requested (can be multiple at once)."""
+        logger = woodchips.get(LOGGER_NAME)
+
         if self.discord:
             Messages.send_discord_message(discord_messages, self.discord_url)
         if self.slack:
@@ -205,4 +232,4 @@ class GithubBug:
         if self.rocketchat:
             Messages.send_rocketchat_message(messages, self.rocketchat_url)
 
-        LOGGER.info(messages)
+        logger.info(messages)